Fix file parsing bug, add new configs for testing
The previous version of the program presumed cost/id etc. values were a
single digit. This caused the program to break if costs were greater or
equal to 10. To fix this, the line was split into an array by space,
rather than accessed by substrings.

To test a 10 node topology, extra config files were added and these
tests were passed.